Created rough functionality of the create new records case

+++ include/menu.h
@@ -6,6 +6,6 @@
 char *getFilename();
 char *getString(int max_length);
 int getInt(int *num);
-int getPrice(float *price);
+int getPrice(double *price);
 
 #endif
